split sae.py into sae_template.py and architectures.py
