+++ saes/utils.py
@@ -2,7 +2,7 @@ import torch
 from EWOthello.data.othello import get
 from EWOthello.mingpt.model import GPTConfig, GPTforProbing
 from EWOthello.mingpt.dataset import CharDataset
-from sae import SAEDummy
+from architectures import SAEDummy
 import random
 from copy import copy
